refactor: moved some logic to sd_card
