+++ core/src/apps/common/sdcard.py
@@ -29,7 +29,7 @@ async def _confirm_retry_wrong_card(ctx: wire.GenericContext) -> None:
         )
 
 
-async def confirm_retry_insert_card(ctx: wire.GenericContext) -> None:
+async def _confirm_retry_insert_card(ctx: wire.GenericContext) -> None:
     if SD_CARD_HOT_SWAPPABLE:
         await confirm_action(
             ctx,
@@ -109,7 +109,7 @@ async def ensure_sdcard(
     from trezor import sdcard
 
     while not sdcard.is_present():
-        await confirm_retry_insert_card(ctx)
+        await _confirm_retry_insert_card(ctx)
 
     if not ensure_filesystem:
         return
@@ -140,6 +140,54 @@ async def ensure_sdcard(
         except OSError:
             # formatting failed, or generic I/O error (SD card power-on failed)
             await confirm_retry_sd(ctx)
+
+
+
+async def ensure_backup_sd_card(
+    ctx: wire.GenericContext, ensure_filesystem: bool = True
+) -> None:
+    """Ensure a SD card is ready for use.
+
+    This function runs the UI flow needed to ask the user to insert a SD card if there
+    isn't one.
+
+    If `ensure_filesystem` is True (the default), it also tries to mount the SD card
+    filesystem, and allows the user to format the card if a filesystem cannot be
+    mounted.
+    """
+    from trezor import sdcard, io
+
+    while not sdcard.is_present():
+        await _confirm_retry_insert_card(ctx)
+
+    if not ensure_filesystem:
+        return
+    fatfs = io.fatfs  # local_cache_attribute
+    while True:
+        try:
+            try:
+                with sdcard.filesystem(mounted=False):
+                    fatfs.mount()
+            except fatfs.NoFilesystem:
+                # card not formatted. proceed out of the except clause
+                pass
+            else:
+                # no error when mounting
+                return
+
+            # Proceed to formatting. Failure is caught by the outside OSError handler
+            with sdcard.filesystem(mounted=False):
+                fatfs.mkfs()
+                fatfs.mount()
+                fatfs.setlabel("TREZOR")
+
+            # format and mount succeeded
+            return
+
+        except OSError:
+            # formatting failed, or generic I/O error (SD card power-on failed)
+            await confirm_retry_sd(ctx)
+
 
 
 async def request_sd_salt(
